Removed old vote aggregator. Renamed VoteAggregatorV2 -> VoteAggregator
